requirements.txt
Flask
google-generativeai
PyMuPDF
gunicorn
python-dotenv
requests
